Refactor functions into localtools
