+++ seed/bluesky/management/commands/migrate_organization.py
@@ -24,63 +24,18 @@ from _localtools import get_static_extradata_mapping_file
 from _localtools import read_building_snapshot_tree_structure
 from _localtools import get_core_organizations
 from _localtools import get_node_sinks
+from _localtools import find_or_create_bluesky_taxlot_associated_with_building_snapshot
+from _localtools import find_or_create_bluesky_property_associated_with_building_snapshot
+from _localtools import load_organization_field_mapping_for_type_exclusions
+from _localtools import load_organization_field_mapping_for_type
+from _localtools import load_organization_property_extra_data_mapping_exclusions
+from _localtools import load_organization_taxlot_extra_data_mapping_exclusions
+from _localtools import load_organization_property_field_mapping
+from _localtools import load_organization_taxlot_field_mapping
+from _localtools import load_raw_mapping_data
 
 logging.basicConfig(level=logging.DEBUG)
 
-
-def find_or_create_bluesky_taxlot_associated_with_building_snapshot(bs, org):
-    desired_field_mapping = load_organization_taxlot_field_mapping(org)
-    reverse_mapping = {y:x for x,y in desired_field_mapping.items()}
-
-    bs_taxlot_val = bs.tax_lot_id
-    if ('jurisdiction_taxlot_identifier' in reverse_mapping and reverse_mapping['jurisdiction_taxlot_identifier'] in bs.extra_data):
-        bs_taxlot_val = bs.extra_data[reverse_mapping['jurisdiction_taxlot_identifier']]
-
-    if bs_taxlot_val is None:
-        tax_lot = seed.bluesky.models.TaxLot(organization=org)
-        tax_lot.save()
-        return tax_lot, True
-
-
-    qry = seed.bluesky.models.TaxLotView.objects.filter(state__jurisdiction_taxlot_identifier=bs_taxlot_val)
-
-    # See if we have any tax lot views that have tax lot states
-    # with that id, if yes, find/return associated property.
-
-    if qry.count():
-        print "HAHA: Found a match!"
-        return qry.first().taxlot, False
-
-    else:
-        tax_lot = seed.bluesky.models.TaxLot(organization=org)
-        tax_lot.save()
-        return tax_lot, True
-
-
-def find_or_create_bluesky_property_associated_with_building_snapshot(bs, org):
-    mapping_field = 'building_portfolio_manager_identifier'
-
-
-    desired_field_mapping = load_organization_property_field_mapping(org)
-    reverse_mapping = {y:x for x,y in desired_field_mapping.items()}
-    bs_property_id = bs.pm_property_id
-    if (mapping_field in reverse_mapping and reverse_mapping[mapping_field] in bs.extra_data):
-        bs_property_id = bs.extra_data[reverse_mapping[mapping_field]]
-
-    if bs_property_id is None:
-        property = seed.bluesky.models.Property(organization=org)
-        property.save()
-        return property, True
-
-    qry = seed.bluesky.models.PropertyView.objects.filter(state__building_portfolio_manager_identifier=bs_property_id)
-
-    if qry.count():
-        print "HOHO: Found a property match"
-        return qry.first().property, False
-    else:
-        property = seed.bluesky.models.Property(organization=org)
-        property.save()
-        return property, True
 
 def copy_extra_data_excluding(extra_data, bad_fields):
     bad_fields = set(bad_fields)
@@ -490,65 +445,6 @@ def create_associated_bluesky_taxlots_properties(org, import_buildingsnapshots, 
                                                                                                                                      property_state_created,
                                                                                                                                      m2m_created))
     return
-
-
-def load_organization_field_mapping_for_type_exclusions(org, type):
-    assert type in ["Tax", "Property"]
-
-    data = load_raw_mapping_data()
-
-    remove_from_extra_data_mapping = []
-
-    for key in data[org]:
-        (table, column) = data[org][key]
-
-        if (table != type):
-            remove_from_extra_data_mapping.append(key)
-
-    return remove_from_extra_data_mapping
-
-
-def load_organization_field_mapping_for_type(org, type):
-    """This returns a list of keys -> (table, attr) to map the key into."""
-    data = load_raw_mapping_data()
-
-    mapping = {}
-    for column in data[org].keys():
-        table, dest_column = data[org][column]
-        if table == type and dest_column != "extra_data":
-            mapping[column] = dest_column
-
-    return mapping
-
-
-def load_organization_property_extra_data_mapping_exclusions(org):
-    return load_organization_field_mapping_for_type_exclusions(org.pk, "Property")
-
-def load_organization_taxlot_extra_data_mapping_exclusions(org):
-    return load_organization_field_mapping_for_type_exclusions(org.pk, "Tax")
-
-def load_organization_property_field_mapping(org):
-    """This returns a list of keys -> (table, attr) to map the key into."""
-    return load_organization_field_mapping_for_type(org.pk, "Property")
-
-def load_organization_taxlot_field_mapping(org):
-    """This returns a list of keys -> (table, attr) to map the key into."""
-    return load_organization_field_mapping_for_type(org.pk, "Tax")
-
-
-def load_raw_mapping_data():
-    # pdb.set_trace()
-    fl = open(get_static_extradata_mapping_file()).readlines()
-
-    fl = filter(lambda x: x.startswith("1,"), fl)
-
-    d = collections.defaultdict(lambda : {})
-    reader = csv.reader(StringIO.StringIO("".join(fl)))
-    for r in reader:
-        org_str, key_name, table, field = r[1:5]
-        d[int(org_str)][key_name] = (table, field)
-
-    return d
